requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
reportlab==4.4.4
psycopg2-binary==2.9.9
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.8.2
